Allow building MAME as shared library.
There is an option to build all the secondary projects as shared libs, but that's not helpful if we need it to be a single dll, and that option (quite wisely) doesn't affect the main binary.

I made kind "ConsoleApp" in main.lua optional. It is overridden by pnacl and winstore*, so I don't touch those 2. Maybe some more configurations need to stay ConsoleApp?

Exporting symbols can be done via MAME_EXPORT macro: extern "C" __declspec(dllexport) (msvc) / __attribute__((visibility("default"))) (gcc)

We pass the object with exported symbols (that aren't necessarily used by the main app) to the linker, instead of making it a part of any static lib. It is a way to ensure that the object is linked in the end and the symbols get exported, otherwise the linker would ignored them. There are other ways to force symbol exporting, but since this isn't meant to be OS dependent, this approach looks better.