remove tls use
I'm not sure if waterbox actually supports this correctly. From disassembly, it appears this produces some fs:0 accesses, which normally is correct for Linux (pointing to pthread_self()), but this is wrong in the context of waterbox, where gs:0x18 is used instead. I'm not sure what this maps to on Windows, I think it's some TLS location, but what TLS would be a mystery. All the info on it just points to its usage for 32 bit windows rather than 64 bit windows.
Technically too I think waterbox is breaking SysV ABI since fs:0 is suppose to point to FsBase (which is also pthread_self()), but that doesn't particularly matter as long as TLS is stayed away from (although I suppose pthread_* calls should still work?)

In practice I'm not sure this actually changes anything. I could suspect https://github.com/TASEmulators/BizHawk/issues/3615 is somehow related, but how it would be would be beyond me (some weird third-party software is involved that Memory and WDude both happen to use? Suppose possible with some odd driver and if one of their hardware pieces matches somewhat?)